stepper: fix NOMORE() sign warning
cant be set to 0xFFFFFFFFUL in the HAL_timers_STM32.h, cause other ones

+++ Marlin/src/module/stepper.cpp
@@ -1298,7 +1298,7 @@ void Stepper::isr() {
     ;
 
     // Limit the value to the maximum possible value of the timer
-    NOMORE(interval, HAL_TIMER_TYPE_MAX);
+    NOMORE(interval, uint32_t(HAL_TIMER_TYPE_MAX));
 
     // Compute the time remaining for the main isr
     nextMainISR -= interval;
